Add files via upload
The file ''Dirichlet-PDE.zip'' is about the program for solving the second-order elliptic PDEs with Dirichlet boundary. The file ''Numman-PDE.zip'' is about the program for solving the second-order elliptic PDEs with Neumann boundary. Both programs employ finite difference numerical methods with the matlab language. The file "'experimenta data'' is for the empiriclal evaluation in the paper.